pyproject.toml
[project]
name = "vision-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-translator>=1.11.4",
    "pillow>=12.1.0",
    "pyobjc>=12.1",
    "pywebview>=6.1",
]
